Add relationship endpoint
